refactor(Content): extract YouTubeUtils
